Removing authorization code to a separated module
